Update recipe paths and pin sphinx version
The underlying structure of mBuild has changed slightly, especially
regarding the location of our default recipes like `Polymer`, etc.

This change was not included in the documentation, which prevented them
from being rendered. This has now been fixed.

Also, sphinx, which is part of the workflow in generating the
documentation we host on the website, underwent some breaking changes
since its 2.0 release. Especially when working with `numpydoc`, many of
the docstrings were mangled and difficult to read. This is now being
pinned within the `conf.py` file located in the `docs` upper level
directory.