Update gitignore for protobuf generated file
Since adding support for protobuf as a way to serialize mb.Compounds,
the `protoc` compiler will generate a python file `compound_pb2.py`.

This file is not to be included in version control, since it is
generated on the user's machine.

This commit adds the `protoc` compiled `compound_pb2.py` file to the
gitignore to minimize confusion for anyone unfamiliar with protobuf.