Remove `nglview` as a hard dependency
Currently we hard require `nglview` for all `mBuild` installs, which
brings along a lot of `ipython` and other `jupyter` dependencies.

These are only ever used when running `ipython` instances and wanting to
visualize a `Compound`. It seems reasonable to migrate this to a soft
dependency, and inform the user if they try to `visualize` that they
need to install `nglview` or `py3Dmol`.

We have not tested if our `run_from_ipython` method returns what we
expect, although I have never ran into an issue where it has even
behaved in an unintended manner. A test has been added to at least check
if an environment is not an ipython env.

This should also minimize the dependencies of `mBuild` for a regular
installation.