Typecast float to int to prevent NaNs in np.arange
In the `SpherePattern` class, `n` is assumed to be an integer value for
the number of points on a sphere. However that was not enforced. This
could lead to floating point math issues with the numpy `arange`
function. See issue
[467](https://github.com/mosdef-hub/mbuild/issues/467) for a more
detailed disucssion.

This patch typecasts `n` to an integer when passing to `numpy.arange`.
This appears to prevent floating point overflow, preventing NaNs for the
generated points on the SpherePattern.