Properly test and import the correct CIF modules (#760)
Thanks to a catch by @uppittu11, the `pycifrw` package was not being
properly tested for an import error, this has now been fixed.

Previously, the `pycifrw` package was conditionally imported using
`import_('pycifrw')`, this is not the module name for `pycifrw`, it is
actually `CifFile`. This change has been made in both the cif reading
logic as well as the tests.

Also, two tests were being skipped always in `test_cif.py`, this was due
to a capitalization error in the `has_garnett` module in
`mbuild.utils.io`.

This has also been fixed.